Fix: by default always keep the order from the server json responses
Undecoding a json encoded dictionnary in python by default will lose the order of the json.

You have to explicitely give a good object_pairs_hook to json dumps in order for it the keep the order of dictionnaries.

clean: also removed (now) useless if

normally, with tags now appropriately used in the "name",
we shouldn't get the same "serie-tags" multiple times in the same answer.

So this test, more specifically the else part, is/was rather useless/odd..