Add tags' names and values for "group by" queries
In queries like select count(value) from measuremente where ... group by * InfluxDB
doesn't set "columns" but the tag names and values are provided by the "tags" raw data.

If tags' dictionary is not empty it uses it as the initial dictionary, so it preserves
the key-values but they will be overriden by columns names... if they exists (not seen yet).