add option to use Requests http(s) proxy
The influxdb-python client does not have support for a web proxy although
the Requests library provides the possibility to use a web proxy for both
http and https requests.
This PR (optionally) adds the feature to use a http(s) proxy by adding
an additional parameter to InfluxDBClient initialization called´"proxies".
This parameter expects that a dict is provided. If this is not set it will
default to { }.
The syntax for the dict is the same as specified in Requests docs:
http://docs.python-requests.org/en/latest/user/advanced/#proxies
proxies = {
  "http": "http://10.10.1.10:3128",
  "https": "http://10.10.1.10:1080",
}